Enabling logging to be turned on in service worker - created a separate class for sw logger - checks local variable if should log or not - created an event to toggle logging remotely
