+++ src/libraries/sw/Log.ts
@@ -1,0 +1,31 @@
+import { OSServiceWorkerFields } from "../../service-worker/types";
+
+declare var self: ServiceWorkerGlobalScope & OSServiceWorkerFields;
+
+export default class Log {
+  static debug(...args: any[]): void {
+    if (!!self.shouldLog) {
+      console.debug(...args);
+    }
+  }
+  static trace(...args: any[]): void {
+    if (!!self.shouldLog) {
+      console.trace(...args);
+    }
+  }
+  static info(...args: any[]): void {
+    if (!!self.shouldLog) {
+      console.info(...args);
+    }
+  }
+  static warn(...args: any[]): void {
+    if (!!self.shouldLog) {
+      console.warn(...args);
+    }
+  }
+  static error(...args: any[]): void {
+    if (!!self.shouldLog) {
+      console.error(...args);
+    }
+  }
+}
